% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_info.R
\name{graph_info}
\alias{graph_info}
\title{graph_info}
\usage{
graph_info(list, logo)
}
\arguments{
\item{list}{A list object that keep all the}

\item{logo}{the magick image object of the logo}
}
\value{
A list has position coordinates for stampping the logo.
}
\description{
Grab width and height of the image. Use these to calculate offset positions for x and y axis to stamp the logo.
}
\examples{
\dontrun{
graph_info(list = images,
           logo_path = "../../")
}
}
\seealso{
Other water_markers: \code{\link{graph_output}},
  \code{\link{graph_read}}, \code{\link{graph_stamp}},
  \code{\link{manage_dir}}
}
\concept{water_markers}
