% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_output.R
\name{graph_output}
\alias{graph_output}
\title{graph_output}
\usage{
graph_output(markered_graphs, output_foler, sitename)
}
\arguments{
\item{markered_graphs}{A list that has all the markered graphs}

\item{output_foler}{A charater string for the output folder name}

\item{sitename}{One of the six sitenames}
}
\value{
A list has nothing in it but graphs should be output into their folders
}
\description{
graph_output
}
\examples{
\dontrun{
graph_output(markered_graphs = ll,
             output_foler = "markered",
             sitename = "Tayler")
}

}
\seealso{
Other water_markers: \code{\link{graph_info}},
  \code{\link{graph_read}}, \code{\link{graph_stamp}},
  \code{\link{manage_dir}}
}
\concept{water_markers}
