% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_read.R
\name{graph_read}
\alias{graph_read}
\title{graph_read}
\usage{
graph_read(sitename = "Tayler", data_source, extension = "*.jpeg$",
  recursive = TRUE, full.names = TRUE)
}
\arguments{
\item{sitename}{farm name}

\item{data_source}{path to the folder holds all sites}

\item{extension}{image file entension}

\item{recursive}{TURE, look depth into subfolder; FALSE, only one level}

\item{full.names}{TURE, default, show the full path for image; FALSE, not recommend, show relative path only}
}
\value{
a list of image objects
}
\description{
Read multiple graphs and then put water markers on.
}
\examples{
\dontrun{
graph_read(sitename = "Tayler",
           data_source,
           extension = "*.jpeg$",
           recursive = TRUE,
           full.names = TRUE)
}
}
\seealso{
Other water_markers: \code{\link{graph_info}},
  \code{\link{graph_output}}, \code{\link{graph_stamp}},
  \code{\link{manage_dir}}
}
\concept{water_markers}
