% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stamp.R
\name{graph_stamp}
\alias{graph_stamp}
\title{graph_stamp}
\usage{
graph_stamp(graphs, logo_info, logo)
}
\arguments{
\item{graphs}{A stack of graphs stored in a list.}

\item{logo_info}{The processed logo magick image object.}

\item{logo}{An adjusted logo image. Must be a magick-image object}
}
\value{
A list of markered graph
}
\description{
Compsite graph with PFR logo.
}
\examples{
\dontrun{
lll <- graph_stamp(graphs = l,
                   logo_info = ll,
                   logo = logo)
}
}
\seealso{
Other water_markers: \code{\link{graph_info}},
  \code{\link{graph_output}}, \code{\link{graph_read}},
  \code{\link{manage_dir}}
}
\concept{water_markers}
