% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_dir.R
\name{manage_dir}
\alias{manage_dir}
\title{manage_dir}
\usage{
manage_dir(data_source = "./", output_loc = "markered")
}
\arguments{
\item{data_source}{source of data}

\item{output_loc}{output location}
}
\value{
logic vector. true for successfuly creat directory(ies); FALSE otherwise
}
\description{
House keeping work.
}
\examples{
\dontrun{
manage_dir(data_souce = "/Analysis/IndividualSiteOutputs/",
           output_loc = "marked_graphes")
}
}
\seealso{
Other water_markers: \code{\link{graph_info}},
  \code{\link{graph_output}}, \code{\link{graph_read}},
  \code{\link{graph_stamp}}
}
\concept{water_markers}
